Move BitNet model to separate file
